requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
supabase
requests
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.13.0
joblib==1.3.2
python-dotenv==1.0.0
psutil
python-dateutil
httpx==0.25.1

# Earth Engine for real satellite data
earthengine-api==1.5.22

# Earth Engine (commented out due to compatibility issues)
# geemap==0.29.6